get rid of tests of plotting options format and rename
